Complete EPC - DT_clf v0.0.ipynb
First version is complete: an unrefined (unpruned, not feature engineered properly, etc.) tree simply predict whether, based on certain inputs/attributes, a certain house would get any improvement recommendations. Considering that only a small fraction of the data contained no improvement recommendations, and the tree (not to mention the data) was poorly optimised, the accuracy of this initial tree is low.